Load movie lists in parallel
When loading movie lists a lot of the time is spent waiting for IO when
loading meta info. When the info for one Movie is received we go on to
the next movie in the list and get meta info for that movie.

This commit will get meta info for all movies in parallel instead. Since
most of the time is spent waiting for a response, we might as well wait
for all of them at once.